requirements.txt
click==7.1.2
Flask==1.1.2
hurry==1.1
hurry.filesize==0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyspeedtest==1.2.7
speedtest-cli==2.1.2
tcp-latency==0.0.8
Werkzeug==1.0.1
